Fix reaction operator and moved some handlers to their correct file
