+++ app/database/operator/server_data_operator/index.ts
@@ -7,6 +7,7 @@ import ChannelHandler, {ChannelHandlerMix} from '@database/operator/server_data_
 import PostHandler, {PostHandlerMix} from '@database/operator/server_data_operator/handlers/post';
 import PostsInChannelHandler, {PostsInChannelHandlerMix} from '@database/operator/server_data_operator/handlers/posts_in_channel';
 import PostsInThreadHandler, {PostsInThreadHandlerMix} from '@database/operator/server_data_operator/handlers/posts_in_thread';
+import ReactionHander, {ReactionHandlerMix} from '@database/operator/server_data_operator/handlers/reaction';
 import TeamHandler, {TeamHandlerMix} from '@database/operator/server_data_operator/handlers/team';
 import UserHandler, {UserHandlerMix} from '@database/operator/server_data_operator/handlers/user';
 import mix from '@utils/mix';
@@ -14,7 +15,7 @@ import mix from '@utils/mix';
 import type {Database} from '@nozbe/watermelondb';
 
 interface ServerDataOperator extends ServerDataOperatorBase, PostHandlerMix, PostsInChannelHandlerMix,
-    PostsInThreadHandlerMix, UserHandlerMix, ChannelHandlerMix, CategoryHandlerMix, TeamHandlerMix {}
+    PostsInThreadHandlerMix, ReactionHandlerMix, UserHandlerMix, ChannelHandlerMix, CategoryHandlerMix, TeamHandlerMix {}
 
 class ServerDataOperator extends mix(ServerDataOperatorBase).with(
     CategoryHandler,
@@ -22,6 +23,7 @@ class ServerDataOperator extends mix(ServerDataOperatorBase).with(
     PostHandler,
     PostsInChannelHandler,
     PostsInThreadHandler,
+    ReactionHander,
     TeamHandler,
     UserHandler,
 ) {
